pyproject.toml
[project]
name = "nldataqueries"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.11",
    "psycopg2-binary>=2.9.10",
    "pypdf>=5.4.0",
    "sentence-transformers>=3.4.1",
    "sqlalchemy>=2.0.39",
    "uuid>=1.30",
    "uvicorn>=0.34.0",
    "weaviate-client>=4.11.2",
]
